# rpc: reduce certificate-related allocations during authentication