@mgartner 2024-11-04
This contributed to 1% of all allocations in a sysbench workload. I was unable to measure an improvement in the sysbench benchmarks added in #133132. But the microbenchmark I added shows the improvement:

```
name                     old time/op    new time/op    delta
Authenticate/tenTen        7.48µs ± 1%    7.65µs ± 1%     ~     (p=0.057 n=4+4)
Authenticate/rootDN        9.72µs ± 2%    9.60µs ± 1%     ~     (p=0.548 n=5+5)
Authenticate/nodeDN        10.2µs ±10%     9.6µs ± 2%     ~     (p=0.310 n=5+5)
Authenticate/commonRoot    7.62µs ± 1%    7.62µs ± 1%     ~     (p=0.841 n=5+5)
Authenticate/commonNode    7.55µs ± 1%    7.70µs ± 1%   +1.94%  (p=0.008 n=5+5)

name                     old alloc/op   new alloc/op   delta
Authenticate/tenTen        57.4kB ± 0%    57.4kB ± 0%     ~     (all equal)
Authenticate/rootDN        57.9kB ± 0%    57.9kB ± 0%     ~     (all equal)
Authenticate/nodeDN        57.9kB ± 0%    57.9kB ± 0%     ~     (all equal)
Authenticate/commonRoot    57.5kB ± 0%    57.4kB ± 0%   -0.11%  (p=0.008 n=5+5)
Authenticate/commonNode    57.5kB ± 0%    57.4kB ± 0%   -0.11%  (p=0.008 n=5+5)

name                     old allocs/op  new allocs/op  delta
Authenticate/tenTen          3.00 ± 0%      3.00 ± 0%     ~     (all equal)
Authenticate/rootDN          28.0 ± 0%      28.0 ± 0%     ~     (all equal)
Authenticate/nodeDN          28.0 ± 0%      28.0 ± 0%     ~     (all equal)
Authenticate/commonRoot      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.008 n=5+5)
Authenticate/commonNode      5.00 ± 0%      3.00 ± 0%  -40.00%  (p=0.008 n=5+5)
```