@tbg 2024-11-05
Do you want this PR to at least mention (but probably fix) https://github.com/cockroachdb/cockroach/issues/133317?
The microbenchmark simply doesn't seem to execute this code path (https://github.com/cockroachdb/cockroach/issues/133317#issuecomment-2457304945), otherwise it would've likely shown up.