@mgartner 2024-11-05
>Haven't fully reviewed yet, just a comment on the benchmark in the first commit. Will give a full review when that's addressed.

The changes are pretty simple if you hide the whitespace changes.