@mgartner 2024-11-06
The new benchmark before/after:

```
name                     old time/op    new time/op    delta
Authenticate/tenTen        42.5ns ± 5%    42.2ns ± 2%      ~     (p=0.794 n=5+5)
Authenticate/rootDN        1.23µs ± 1%    1.21µs ± 0%    -1.85%  (p=0.008 n=5+5)
Authenticate/nodeDN        1.23µs ± 1%    1.21µs ± 0%    -1.24%  (p=0.008 n=5+5)
Authenticate/commonRoot     127ns ± 3%      76ns ± 0%   -39.82%  (p=0.008 n=5+5)
Authenticate/commonNode     129ns ± 4%      76ns ± 0%   -40.70%  (p=0.008 n=5+5)

name                     old alloc/op   new alloc/op   delta
Authenticate/tenTen         0.00B          0.00B           ~     (all equal)
Authenticate/rootDN          464B ± 0%      464B ± 0%      ~     (all equal)
Authenticate/nodeDN          464B ± 0%      464B ± 0%      ~     (all equal)
Authenticate/commonRoot     64.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
Authenticate/commonNode     64.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)

name                     old allocs/op  new allocs/op  delta
Authenticate/tenTen          0.00           0.00           ~     (all equal)
Authenticate/rootDN          25.0 ± 0%      25.0 ± 0%      ~     (all equal)
Authenticate/nodeDN          25.0 ± 0%      25.0 ± 0%      ~     (all equal)
Authenticate/commonRoot      2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
Authenticate/commonNode      2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
```